explicitely cast

+++ geometry/r3_element_body.hpp
@@ -315,7 +315,9 @@ quantities::Product<LScalar, RScalar> Dot(
 inline R3Element<double> BasisVector(int const i) {
   DCHECK_GE(i, 0) << i;
   DCHECK_LT(i, 3) << i;
-  return {i == 0, i == 1, i == 2};
+  return {static_cast<double>(i == 0),
+          static_cast<double>(i == 1),
+          static_cast<double>(i == 2)};
 }
 
 }  // namespace geometry
